添加反向交割撮合数据

+++ testCase/ContractTestCase/07_Noti/ws_trade/TestContractNoti_ws_trade_017.py
@@ -63,7 +63,7 @@ class TestContractNoti_ws_trade_017:
                 assert False
             if tradedetail['direction'] == None:
                 assert False
-            if int(tradedetail['price']) == int(self.buy_price):
+            if float(tradedetail['price']) == float(self.buy_price):
                 assert False
             if tradedetail['quantity'] == None:
                 assert False
